recent pms reducer: Add `EVENT_NEW_MESSAGE`.
Since the `PmConversationsCard` needs to be updated when a new
message comes in and the data it uses to be updated is in
`recentPrivateConversations` key in the Redux state, this event
catch adds/amends the data in the state to correctly match the most
recent private messages.

A unit test file has been created for this use-case as well. It is
`strict-local`, and uses `exampleData.js` for its data source.

[ray: rewrote test file to use new functions; added self-user ID to
  tests.]