api: send 'Accept-Language' header
The Zulip server relies on the `Accept-Language` header to provide
localized error messages in its responses.

Unfortunately, we need the user's current locale to send that header,
and there's no good path from Redux into here. So, for now, we use a
bad one (accompanied by plans for its removal in the future).