heartbeat [nfc]: Add fine-grained event tracking to tests.
We don't really have any tests, so far, acting as evidence that
heartbeats don't fundamentally misbehave.

In preparation for writing some (and for defining "fundamentally
misbehave"), add state-change tracking and callback tracking to the
heartbeat tests.

Note that, due to the mock timers we're using, the tracking functions
added in this commit are subtly incorrect! This will be elaborated on
in the following commit, in which a correct version will be
substituted.