XXXXXXXXXX begin: react-native 0.60.0 upgrade XXXXXXXXXX
The following series of commits applies, in pieces, the various
changes needed to upgrade React Native to version 0.60.0.

Unfortunately, the recommended way to upgrade React Native [0] is
literally to apply the diffs between the two versions' outputs for
`react-native init`. This is clearly not feasible to do automatically
for a project of any interesting complexity or level of customization,
which leaves us in the unpleasant situation of needing to do it
manually.

The React Native upgrade instructions include a link to an "Upgrade
Helper" tool, which presents the diff much as it would be on
GitHub:

  https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.0

This tool is being *very* actively developed; a description of its
features at time of writing is expected to be out of date by the time
this commit message is read.

Reviewers, and archaeologists, should note in particular that any
comments it includes inline may not have been present at time of
writing. (Those that were will be mentioned as they become relevant.)

[0] https://facebook.github.io/react-native/docs/upgrading

------------------------------------------------------------------------

React Native 0.60 comes with many significant changes to the build
process. Applying them all at once would leave us with an opaque
commit and, if fully explained, a monstrously-long commit message of
questionable utility.

Unfortunately, few meaningful or useful subsets of these changes are
functional individually. It is therefore broken into nonfunctional
pieces, which are relatively-easily reviewable but completely
unsuitable to make part of the mainline git history.

(Exactly how these changes will end up in the repo is not yet decided.
The top candidate so far is `squash + ref to tag`.)