tests: work around timer bug in node
A followup to #3763, and not at all in the way I was expecting.

Node.js can sometimes return slightly early -- or at least, "early",
as perceived by `Date.now()`. As this appears to be due only to
rounding issues, we hack around this by subtracting 1 from the minimum
expected value.