xxx rn-0.60: move React Native core components to Podfile
Move all React Native required core components to the Podfile.  This
commit is largely a manual application of the 0.59-to-0.60 diff for
`project.pbxproj` and `ios/Podfile`.

The difference is believed to consist entirely of:
  a) those tvOS build fragments in the project file which could not
easily be included due to the prior removal of the entire structural
subtree in which they were located; and
  b) the few tvOS-relevant lines in the Podfile (which have been
commented out).

The remaining tvOS-only fragments will be removed by Xcode in a
following commit.