xxx rn-0.60: `(cd ios; pod install)`
Commit the result of running `(cd ios; pod install)`. The expected
component of this commit is the modifications to Podfile.lock (which
may be thought of as yarn.lock for Podfiles).

In addition to this, Cocoapods injects a new build phase to copy
files, at the behest of `react-native-vector-icons`.  (As the initial
"[CP]" denotes, this build phase is Cocoapods-controlled.)

At this point, all necessary nodifications have been made for the iOS
build to succeed. (Confirmed in debug and release.)