XXX outbox: Add error-checking to EVENT_NEW_MESSAGE handling.
If an EVENT_NEW_MESSAGE comes in with an unknown `local_message_id`,
report an error to Sentry. (This is not a functional change.)

If it is known, but the message it identifies in not in the sent
state, also report an error, and leave the message in place.

This is technically a functional change! In the second case, we now
leave the identified message in place, instead of returning it.

XXX

Under
normal operation, this is technically possible, but will only occur in
the unlikely case that a message is blocked from being sent until
*just* before it decays. XXX should be very rare

XXX mostly chosen because easier to document