recent convs: Normalize and retain user IDs received from server.
The `user_id`s in `recent_private_conversations` do not include the
self-user. As there are enough distinct ways to express user-lists
already, we instead adjust this to match `Message#display_recipient`,
which always includes the self-user (and is the API data source we'll
most commonly be using alongside this data, anyway).

Archaeological note: the user IDs in `recent_private_conversations`
were not sorted until 2.2-dev-53-g405a529340 (zulip/zulip@405a529340);
this was backported to the 2.1 branch as 2.1.1-50-gd452ad31e0
(zulip/zulip@d452ad31e0).