recent convs: Add new implementation of getRecentConversations.
Add new implementation of `getRecentConversations`, based on new
`recent_private_conversations` data.

We can't get this data on pre-2.1 servers, so we continue using the
legacy data there, to avoid a regression for users on such servers.
(Without this, they would never see any old conversations in the PMs
window; only those with messages that have come in during the current
session would be visible.)

In upcoming commits:
  * New tests for the new implementation (alongside the retained tests
    for the legacy implementation).
  * Adjusting the main selector to be more efficient, rather than
    potentially computing both the modern and legacy implementations
    on every call.